feat: dev build script can build only packages

+++ build/build.sh
@@ -6,17 +6,21 @@
 
 export CI_BUILD="no"
 export SHOULD_BUILD="yes"
+export SKIP_BUILD="no"
 export SKIP_PACKAGES="yes"
 export VSCODE_LATEST="no"
 export VSCODE_QUALITY="stable"
 
-while getopts ":ilp" opt; do
+while getopts ":ilop" opt; do
   case "$opt" in
     i)
       export VSCODE_QUALITY="insider"
       ;;
     l)
       export VSCODE_LATEST="yes"
+      ;;
+    o)
+      export SKIP_BUILD="yes"
       ;;
     p)
       export SKIP_PACKAGES="no"
@@ -45,18 +49,35 @@ else
 fi
 
 echo "OS_NAME=\"${OS_NAME}\""
+echo "SKIP_BUILD=\"${SKIP_BUILD}\""
 echo "SKIP_PACKAGES=\"${SKIP_PACKAGES}\""
 echo "VSCODE_ARCH=\"${VSCODE_ARCH}\""
 echo "VSCODE_LATEST=\"${VSCODE_LATEST}\""
 echo "VSCODE_QUALITY=\"${VSCODE_QUALITY}\""
 
-rm -rf vscode* VSCode*
+if [[ "${SKIP_BUILD}" == "no" ]]; then
+  rm -rf vscode* VSCode*
 
-. get_repo.sh
-. build.sh
+  . get_repo.sh
 
-if [[ "${VSCODE_QUALITY}" == "insider" && "${VSCODE_LATEST}" == "yes" ]]; then
-  echo "$( cat "insider.json" | jq --arg 'tag' "${MS_TAG/\-insider/}" --arg 'commit' "${MS_COMMIT}" '. | .tag=$tag | .commit=$commit' )" > "insider.json"
+  # save variables for later
+  echo "MS_TAG=\"${MS_TAG}\"" > build.env
+  echo "MS_COMMIT=\"${MS_COMMIT}\"" >> build.env
+  echo "RELEASE_VERSION=\"${RELEASE_VERSION}\"" >> build.env
+  echo "BUILD_SOURCEVERSION=\"${BUILD_SOURCEVERSION}\"" >> build.env
+
+  . build.sh
+
+  if [[ "${VSCODE_QUALITY}" == "insider" && "${VSCODE_LATEST}" == "yes" ]]; then
+    echo "$( cat "insider.json" | jq --arg 'tag' "${MS_TAG/\-insider/}" --arg 'commit' "${MS_COMMIT}" '. | .tag=$tag | .commit=$commit' )" > "insider.json"
+  fi
+else
+  . build.env
+
+  echo "MS_TAG=\"${MS_TAG}\""
+  echo "MS_COMMIT=\"${MS_COMMIT}\""
+  echo "RELEASE_VERSION=\"${RELEASE_VERSION}\""
+  echo "BUILD_SOURCEVERSION=\"${BUILD_SOURCEVERSION}\""
 fi
 
 if [[ "${SKIP_PACKAGES}" == "no" ]]; then
